requirements.txt
fastapi
uvicorn
python-dotenv
requests
sqlalchemy
mysqlclient
mysql-connector-python
pydantic
datetime